Replace HttpWebRequest/Response with HttpClient; make Get() async,
I'm experiencing a very slight delay (0.1 sec?) when clicking Button_GetDevice_Location. After a few rabbit trails:

Kicked HttpWebRequest and HttpWebResponse to the curb, replacing with HttpClient.
Get() now directly returns StreamReader.ReadToEnd().

Get() is now async (or should be, anyway) with the intention of preventing deadlock (that is, the UI from going unresponsive). Seems to work so far?

I'll be honest, I'm not entirely sure if any of this is "best practice" or "best solution" but nothing's broken yet, lol.