Publish/Release Build Test
While running Vigil is Visual Studio, in Debug build, I was experiencing a delay between the first button click and the actions executing. So, I published the project and ran the Release build exe, and there is no delay.

Was not expecting Visual Studio to create so many new files. I gotta clean this up.