Pin Movement Implemented
A placeholder "pin" now moves to the appropriate spot on a crude map, based on device location as reported by the server. Everything is currently hardcoded, but eventually will be user-settable.

For some reason though, with each update the pin animation "jumps" back to where it started (0,0) to move back to where it should be. Interesting bug.