Update Pin Movement
Pin doesn't move via animation anymore, but it does move appropriately according to the device's updated location.

I'm not sure how important animation is, so it may or may not be reimplemented.

Next step is to begin implementing User-Settable values (that is, user picks the image, user sets location coordinates, etc.)